a3

+++ pybricks/pyBricks1/main.py
@@ -109,7 +109,7 @@ async def get_the_vegetables():
     await WheelController.move_wheels_forward_in_straight_line(float(180))
     await WheelController.wheel_left_turn()
     await multitask(GripperController.release_element_using_both_arms(),
-                    WheelController.move_wheels_forward_in_straight_line(float(180), float(100)))
+                    WheelController.move_wheels_forward_in_straight_line(float(185), float(100)))
     await GripperController.grip_element_using_both_arms()
 
     # going long straight to the compose area
@@ -156,7 +156,7 @@ async def water_the_green_plants_and_move_decay_plants():
     await WheelController.move_wheels_forward_in_straight_line(float(210))
     await multitask(GripperController.grip_element_using_both_arms(),
                     WheelController.wheel_right_turn_with_angle(float(20)))
-    await WheelController.move_wheels_forward_in_straight_line(float(310))
+    await WheelController.move_wheels_forward_in_straight_line(float(305))
     await WheelController.wheel_left_turn()
     await WheelController.move_wheels_forward_in_straight_line(float(60))
     await WheelController.move_wheels_towards_element_then_stop_at_marker()
